veo_requirements.txt
Flask>=2.3
Pillow>=10.0
